+++ _examples/logger.go
@@ -29,8 +29,10 @@ import (
 
 func main() {
 
-    // NewStdoutLogger creates a new Logger holder to standard output, generally a terminal or a console.
-    logger := logit.NewStdoutLogger(logit.DebugLevel)
+    // NewDevelopLogger creates a new Logger holder for developing, generally log to terminal or console.
+    // You can switch to logit.NewProductionLogger for production environment.
+    //logger := logit.NewProductionLogger(os.Stdout)
+    logger := logit.NewDevelopLogger()
 
     // Then you will be easy to log!
     logger.Debug("this is a debug message!")
@@ -38,15 +40,23 @@ func main() {
     logger.Warn("this is a warn message!")
     logger.Error("this is an error message!")
 
-    // NewLogger creates a new Logger holder.
-    // The first parameter "os.Stdout" is a writer for logging.
-    // As you know, file also can be written, just replace "os.Stdout" with your file!
-    // The second parameter "logit.DebugLevel" is the level of this Logger.
-    logger = logit.NewLogger(os.Stdout, logit.DebugLevel)
-
-    // If you want format your time, try this:
-    logger.SetFormatOfTime("2006/01/02 15:04:05")
+    // NewLogger creates a new Logger holder with given level and handlers.
+    // As you know, file also can be written, just replace os.Stdout with your file!
+    // A logger is made of level and handlers, so we provide some handlers for use, see logit.Handler.
+    // This method is the most original way to create a logger for use.
+    logger = logit.NewLogger(logit.DebugLevel, logit.NewDefaultHandler(os.Stdout, logit.NewDefaultEncoder("2006/01/02 15:04:05")))
     logger.Info("What time is it now?")
+
+    // NewLoggerFrom creates a new Logger holder with given config.
+    // The config has all the things to create a logger, such as level.
+    // We provide some encoders: default encoder and json encoder.
+    // See logit.Encoder to check more information.
+    logger = logit.NewLoggerFrom(logit.Config{
+        Level:   logit.DebugLevel,
+        Writer:  os.Stdout,
+        Encoder: logit.NewJsonEncoder("2006/01/02 15:04:05", false),
+    })
+    logger.Info("I am a json log!")
 
     // If you want to output log with file info, try this:
     logger.EnableFileInfo()
@@ -60,7 +70,4 @@ func main() {
         r := rand.New(rand.NewSource(time.Now().Unix()))
         return "debug rand int: " + strconv.Itoa(r.Intn(100))
     })
-
-    // If you want to change logger's writer, try this:
-    logger.ChangeWriterTo(os.Stdout)
 }