+++ logger.go
@@ -19,7 +19,6 @@
 package logit
 
 import (
-    "io"
     "runtime"
     "strconv"
     "sync"
@@ -29,18 +28,6 @@ import (
 // Logger is a struct to log.
 type Logger struct {
 
-    // writer is the output of this Logger.
-    writer io.Writer
-
-    // handlers is the slice of log handlers.
-    // You can add your handler for some situations.
-    // See LoggerHandler.
-    handlers []LoggerHandler
-
-    // formatOfTime is the format for formatting time.
-    // Default is "2006-01-02 15:04:05", see DefaultFormatOfTime.
-    formatOfTime string
-
     // level is the level representation of the Logger.
     // In this version of logit, there are four levels:
     //
@@ -49,76 +36,67 @@ type Logger struct {
     // The righter level has higher visibility which means
     // one debug message will not be logged in one Logger in InfoLevel.
     // That's we called level-based logging.
-    level LoggerLevel
+    level Level
 
-    // running is the status of the Logger.
-    // true means the Logger is running, false means the Logger is shutdown.
-    running bool
+    // handlers is the slice of log handlers.
+    // You can add your handler for some situations.
+    // See LoggerHandler.
+    handlers []Handler
 
     // needFileInfo is a flag to check if msg should contain file info.
     // This step is useful but too expensive, so default is false.
     needFileInfo bool
 
+    // logs is an object pool cache some Log holder.
+    // Use a pool is for reducing memory allocation.
+    logs *sync.Pool
+
     // mu is for safe concurrency.
-    mu sync.RWMutex
+    mu *sync.RWMutex
 }
 
-// DefaultFormatOfTime is the default format for formatting time.
-const DefaultFormatOfTime = "2006-01-02 15:04:05"
-
-// NewLogger creates one Logger with given out and level.
-// The first parameter writer is the writer for logging.
-// The second parameter level is the level of this Logger.
-// It returns a new running Logger holder.
-func NewLogger(writer io.Writer, level LoggerLevel) *Logger {
-    return NewLoggerWithHandlers(writer, level, DefaultLoggerHandler)
-}
-
-// NewLoggerWithHandlers creates one Logger with given out and level and handlers.
+// NewLogger creates one Logger with given out and level and handlers.
 // The first parameter writer is the writer for logging.
 // The second parameter level is the level of this Logger.
 // The third parameter handlers is all logger handlers for handling each log.
 // It returns a new running Logger holder.
-func NewLoggerWithHandlers(writer io.Writer, level LoggerLevel, handlers ...LoggerHandler) *Logger {
+func NewLogger(level Level, handlers ...Handler) *Logger {
 
     // 至少添加一个日志处理器，否则直接报错
     if len(handlers) < 1 {
         panic("You must add at least one handler!")
     }
 
-    return &Logger{
-        writer:       writer,
-        formatOfTime: DefaultFormatOfTime,
-        handlers:     handlers,
+    // 创建 logger 对象
+    logger := &Logger{
         level:        level,
-        running:      true,
+        handlers:     handlers,
         needFileInfo: false,
+        mu:           &sync.RWMutex{},
     }
-}
 
-// Enable sets l on running status.
-func (l *Logger) Enable() {
-    l.mu.Lock()
-    defer l.mu.Unlock()
-    l.running = true
-}
+    // 初始化 logs 对象池
+    logger.logs = &sync.Pool{
+        New: func() interface{} {
+            return &Log{
+                logger: logger,
+                extra:  map[string]string{},
+            }
+        },
+    }
 
-// Disable sets l on shutdown status.
-func (l *Logger) Disable() {
-    l.mu.Lock()
-    defer l.mu.Unlock()
-    l.running = false
+    return logger
 }
 
 // ChangeLevelTo will change the level of current Logger to newLevel.
-func (l *Logger) ChangeLevelTo(newLevel LoggerLevel) {
+func (l *Logger) ChangeLevelTo(newLevel Level) {
     l.mu.Lock()
     defer l.mu.Unlock()
     l.level = newLevel
 }
 
 // Level returns the logger level of l.
-func (l *Logger) Level() LoggerLevel {
+func (l *Logger) Level() Level {
     l.mu.RLock()
     defer l.mu.RUnlock()
     return l.level
@@ -143,8 +121,8 @@ func (l *Logger) DisableFileInfo() {
 
 // AddHandlers adds more handlers to l, and all handlers added before will be retained.
 // If you want to remove all handlers, try l.SetHandlers().
-// See logit.DefaultLoggerHandler.
-func (l *Logger) AddHandlers(handlers ...LoggerHandler) {
+// See logit.Handler.
+func (l *Logger) AddHandlers(handlers ...Handler) {
     l.mu.Lock()
     defer l.mu.Unlock()
     l.handlers = append(l.handlers, handlers...)
@@ -154,8 +132,8 @@ func (l *Logger) AddHandlers(handlers ...LoggerHandler) {
 // If you want to add more handlers rather than replace them, try l.AddHandlers.
 // Notice that at least one handler should be added, so if len(handlers) < 1, it returns false
 // which means setting failed. Return true if setting is successful.
-// See logit.DefaultLoggerHandler.
-func (l *Logger) SetHandlers(handlers ...LoggerHandler) bool {
+// See logit.Handler.
+func (l *Logger) SetHandlers(handlers ...Handler) bool {
 
     // 必须添加至少一个处理器
     if len(handlers) < 1 {
@@ -171,42 +149,29 @@ func (l *Logger) SetHandlers(handlers ...LoggerHandler) bool {
 }
 
 // Handlers returns all handlers of l in a copy slice.
-func (l *Logger) Handlers() []LoggerHandler {
+func (l *Logger) Handlers() []Handler {
     l.mu.RLock()
     defer l.mu.RUnlock()
 
     // 返回的是日志处理器的副本，防止被非法篡改
-    handlers := make([]LoggerHandler, 0, len(l.handlers))
+    handlers := make([]Handler, 0, len(l.handlers))
     return append(handlers, l.handlers...)
 }
 
-// SetFormatOfTime sets format of time as you want.
-// Default is "2006-01-02 15:04:05", see DefaultFormatOfTime.
-func (l *Logger) SetFormatOfTime(formatOfTime string) {
-    l.mu.Lock()
-    defer l.mu.Unlock()
-    l.formatOfTime = formatOfTime
+// newLog returns a Log holder from object pool.
+// Notice that not every holder returned is new, as you know, that is why we use a pool.
+func (l *Logger) newLog(level Level, msg string) *Log {
+    log := l.logs.Get().(*Log)
+    log.level = level
+    log.now = time.Now()
+    log.msg = msg
+    return log
 }
 
-// FormatOfTime returns the format of time in l.
-func (l *Logger) FormatOfTime() string {
-    l.mu.RLock()
-    defer l.mu.RUnlock()
-    return l.formatOfTime
-}
-
-// ChangeWriterTo changes current writer to newWriter.
-func (l *Logger) ChangeWriterTo(newWriter io.Writer) {
-    l.mu.Lock()
-    defer l.mu.Unlock()
-    l.writer = newWriter
-}
-
-// Writer returns the writer of l.
-func (l *Logger) Writer() io.Writer {
-    l.mu.RLock()
-    defer l.mu.RUnlock()
-    return l.writer
+// releaseLog releases log to object pool so that this log can be reused next time.
+func (l *Logger) releaseLog(log *Log) {
+    log.extra = map[string]string{}
+    l.logs.Put(log)
 }
 
 // callDepth is the depth of the method calling stack, which is about file name and line.
@@ -214,15 +179,13 @@ const callDepth = 3
 
 // log can output msg to l.writer, notices that level will affect the visibility of this msg.
 // Notice that callDepth is caller sensitive, and the value is about file name and line.
-func (l *Logger) log(callDepth int, level LoggerLevel, msg string) {
+func (l *Logger) log(callDepth int, level Level, msg string) {
 
     // 加上读锁
     l.mu.RLock()
 
-    // 以下两种条件直接返回，不记录日志：
-    // 1. 日志处于禁用状态，也就是 l.running = false
-    // 2. 日志记录器的日志级别高于这条记录的日志级别
-    if !l.running || l.level > level {
+    // 日志记录器的级别高于日志的级别，不进行记录：
+    if l.level > level {
         l.mu.RUnlock()
         return
     }
@@ -235,38 +198,42 @@ func (l *Logger) log(callDepth int, level LoggerLevel, msg string) {
     needFileInfo := l.needFileInfo
     l.mu.RUnlock()
 
+    // 处理日志
+    log := l.newLog(level, msg)
+    defer l.releaseLog(log)
+
     // 如果需要文件信息，对当前的 msg 进行包装
     if needFileInfo {
-        msg = wrapMessageWithFileInfo(callDepth, msg)
+        wrapLogWithFileInfo(callDepth, log)
     }
-
-    // 处理日志
-    l.handleLog(level, time.Now(), msg)
+    l.handleLog(log)
 }
 
 // handleLog handles log with l.handlers.
 // Notice that if one handler returns false, then all handlers after it
 // will not use anymore.
-func (l *Logger) handleLog(level LoggerLevel, now time.Time, msg string) {
+func (l *Logger) handleLog(log *Log) {
     for _, handler := range l.handlers {
-        if !handler.handle(l, level, now, msg) {
-            break
+        if !handler.Handle(log) {
+            return
         }
     }
 }
 
-// wrapMessageWithFileInfo wraps msg with file info.
+// wrapLogWithFileInfo wraps log with file info.
 // This function is too expensive because of runtime.Caller.
 // Notice that callDepth is the depth of calling stack. See callDepth.
-func wrapMessageWithFileInfo(callDepth int, msg string) string {
+func wrapLogWithFileInfo(callDepth int, log *Log) {
 
     // 这个 callDepth 是 runtime.Caller 方法的参数，表示上面第几层调用者信息
     _, file, line, ok := runtime.Caller(callDepth)
     if !ok {
-        return "[unknown file:unknown line] " + msg
+        log.extra["file"] = "unknown file"
+        log.extra["line"] = "-1"
     }
 
-    return "[" + file + ":" + strconv.Itoa(line) + "] " + msg
+    log.extra["file"] = file
+    log.extra["line"] = strconv.Itoa(line)
 }
 
 // Debug will output msg as a debug message.