+++ logger_test.go
@@ -25,56 +25,55 @@ import (
     "strings"
     "sync"
     "testing"
-    "time"
 )
 
 // 测试日志记录器的 Debug 方法
 func TestLoggerDebug(t *testing.T) {
-    logger := NewLogger(os.Stdout, DebugLevel)
+    logger := NewLogger(DebugLevel, NewDefaultHandler(os.Stdout, NewDefaultEncoder(DefaultTimeFormat)))
     logger.Debug("这是 debug 信息。。。")
 }
 
 // 测试日志记录器的 Debug 方法
 func TestLoggerInfo(t *testing.T) {
-    logger := NewLogger(os.Stdout, DebugLevel)
+    logger := NewLogger(DebugLevel, NewDefaultHandler(os.Stdout, NewDefaultEncoder(DefaultTimeFormat)))
     logger.Info("这是 info 信息。。。")
 }
 
 // 测试日志记录器的 Debug 方法
 func TestLoggerWarn(t *testing.T) {
-    logger := NewLogger(os.Stdout, DebugLevel)
+    logger := NewLogger(DebugLevel, NewDefaultHandler(os.Stdout, NewDefaultEncoder(DefaultTimeFormat)))
     logger.Warn("这是 warn 信息。。。")
 }
 
 // 测试日志记录器的 Debug 方法
 func TestLoggerError(t *testing.T) {
-    logger := NewLogger(os.Stdout, DebugLevel)
+    logger := NewLogger(DebugLevel, NewDefaultHandler(os.Stdout, NewDefaultEncoder(DefaultTimeFormat)))
     logger.Error("这是 error 信息。。。")
 }
 
 // 测试调整日志记录器为运行状态的方法
 func TestLoggerEnable(t *testing.T) {
-    logger := NewLogger(os.Stdout, DebugLevel)
+    logger := NewLogger(DebugLevel, NewDefaultHandler(os.Stdout, NewDefaultEncoder(DefaultTimeFormat)))
     logger.Debug("1. 这是 debug 信息。。。")
-    logger.Disable()
+    logger.ChangeLevelTo(OffLevel)
     logger.Debug("2. 这是 debug 信息。。。")
-    logger.Enable()
+    logger.ChangeLevelTo(DebugLevel)
     logger.Debug("3. 这是 debug 信息。。。")
 }
 
 // 测试日志记录器的级别控制是否可用
-func TestLoggerLevel(t *testing.T) {
-    logger := NewLogger(os.Stdout, WarnLevel)
-    logger.Warn(logger.Level().String())
+func TestLevel(t *testing.T) {
+    logger := NewLogger(DebugLevel, NewDefaultHandler(os.Stdout, NewDefaultEncoder(DefaultTimeFormat)))
+    logger.Debug(logger.Level().String())
     logger.Info("这条 info 级别的内容可以显示吗？")
-    logger.Error("这条 error 级别的内容可以显示吗？")
     logger.Warn("这条 warn 级别的内容可以显示吗？")
+    logger.Error("这条 error 级别的内容可以显示吗？")
 }
 
 // 测试更改日志级别是否可用
 func TestLoggerChangeLevelTo(t *testing.T) {
-    logger := NewLogger(os.Stdout, WarnLevel)
-    logger.Info("Log level is warn, so info message will not be logged!")
+    logger := NewLogger(WarnLevel, NewDefaultHandler(os.Stdout, NewDefaultEncoder(DefaultTimeFormat)))
+    logger.Info("Logger's level is warn, so info message will not be logged!")
 
     logger.ChangeLevelTo(InfoLevel)
     logger.Info("Now info message will be logged!")
@@ -83,22 +82,9 @@ func TestLoggerChangeLevelTo(t *testing.T) {
     logger.Warn("Now only error messages will be logged!")
 }
 
-// 测试使用日志处理器创建日志记录器的方法
-func TestNewLoggerWithHandlers(t *testing.T) {
-
-    defer func() {
-        err := recover()
-        if err == nil {
-            t.Fatal("没有传入日志处理器，应该报错的，但是没有报！")
-        }
-    }()
-
-    NewLoggerWithHandlers(os.Stdout, DebugLevel)
-}
-
 // 测试文件信息显示的开关是否可用
 func TestLoggerEnableAndDisableFileInfo(t *testing.T) {
-    logger := NewLogger(os.Stdout, WarnLevel)
+    logger := NewLogger(DebugLevel, NewDefaultHandler(os.Stdout, NewDefaultEncoder(DefaultTimeFormat)))
     logger.Warn("没有文件信息！")
     logger.EnableFileInfo()
     logger.Warn("有文件信息？是否正确？")
@@ -106,43 +92,46 @@ func TestLoggerEnableAndDisableFileInfo(t *testing.T) {
     logger.Warn("现在应该没有文件信息了吧！")
 }
 
+type myHandler struct{}
+
+// Customize your own handler.
+func (mh *myHandler) Handle(log *Log) bool {
+    os.Stdout.WriteString("handler1: " + log.Msg() + "\n")
+    return true
+}
+
 // 测试增加处理器是否可用
 func TestLoggerAddHandlersAndSetHandlers(t *testing.T) {
-    logger := NewLogger(os.Stdout, InfoLevel)
+    logger := NewLogger(DebugLevel, NewDefaultHandler(os.Stdout, NewDefaultEncoder(DefaultTimeFormat)))
     logger.Info("当前的日志处理器：" + fmt.Sprintf("%v", logger.handlers))
 
-    handlers1 := func(logger *Logger, level LoggerLevel, now time.Time, msg string) bool {
-        logger.writer.Write([]byte("第一个日志处理器！\n"))
-        return true
+    if len(logger.handlers) != 1 {
+        t.Fatal("处理器个数不正确！")
     }
 
-    handlers2 := func(logger *Logger, level LoggerLevel, now time.Time, msg string) bool {
-        logger.writer.Write([]byte("第二个日志处理器！\n"))
-        return true
-    }
-
-    logger.AddHandlers(handlers1, handlers2)
+    logger.AddHandlers(&myHandler{})
     logger.Info("当前的日志处理器：" + fmt.Sprintf("%v", logger.handlers))
+
+    if len(logger.handlers) != 2 {
+        t.Fatal("处理器个数不正确！")
+    }
 
     ok := logger.SetHandlers()
     if ok {
         t.Fatal("SetHandlers 应该返回 false！")
     }
-    logger.SetHandlers(handlers1, handlers2)
-    logger.Info("当前的日志处理器：" + fmt.Sprintf("%v", logger.handlers))
-}
 
-// 测试更改时间格式化标准的方法
-func TestLoggerSetFormatOfTime(t *testing.T) {
-    logger := NewLogger(os.Stdout, InfoLevel)
-    logger.Info("当前时间格式化信息！" + logger.FormatOfTime())
-    logger.SetFormatOfTime("2006年01月02日 15点04分05秒")
-    logger.Info("更改之后的时间格式化信息！" + logger.FormatOfTime())
+    logger.SetHandlers(&myHandler{})
+    logger.Info("当前的日志处理器：" + fmt.Sprintf("%v", logger.handlers))
+
+    if len(logger.handlers) != 1 {
+        t.Fatal("处理器个数不正确！")
+    }
 }
 
 // 测试获取日志处理器的方法
 func TestLoggerHandlers(t *testing.T) {
-    logger := NewStdoutLogger(InfoLevel)
+    logger := NewDevelopLogger()
     handlers := logger.Handlers()
 
     // 需要判断地址是否一样，一样说明有问题
@@ -161,30 +150,14 @@ func TestLoggerHandlers(t *testing.T) {
     })
 
     // 尝试非法篡改日志处理器
-    handlers[0] = func(logger *Logger, level LoggerLevel, now time.Time, msg string) bool {
-        fmt.Println("哈哈哈被非法篡改了！！！")
-        return false
-    }
+    handlers[0] = &myHandler{}
     logger.Warn("这条信息需要显示出来！")
-}
-
-// 测试更改写出器的方法
-func TestLoggerChangeWriterTo(t *testing.T) {
-    logger := NewStdoutLogger(DebugLevel)
-    fmt.Println(logger.writer)
-    logger.Debug("哈哈哈")
-
-    file, _ := os.Create("Z:/TestLoggerChangeWriterTo.log")
-    logger.ChangeWriterTo(file)
-    fmt.Println(logger.writer)
-    logger.Debug("哈哈哈文件？")
 }
 
 // 测试并发情况下使用 Logger
 func TestLoggerInConcurrency(t *testing.T) {
 
-    logger := NewFileLogger("Z:/TestLoggerInConcurrency.log", DebugLevel)
-    //logger := NewStdoutLogger(DebugLevel)
+    logger := NewLogger(DebugLevel, NewDefaultHandler(os.Stdout, NewDefaultEncoder(DefaultTimeFormat)))
 
     group := sync.WaitGroup{}
     for i := 0; i < 100; i++ {
@@ -194,9 +167,6 @@ func TestLoggerInConcurrency(t *testing.T) {
                 logger.ChangeLevelTo(InfoLevel)
             }
             logger.Info(strconv.Itoa(num))
-            if num == 60 || num == 90 {
-                logger.ChangeWriterTo(os.Stdout)
-            }
             group.Done()
         }(i)
     }